+++ src/main/java/wooteco/subway/domain/member/Favorite.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain.member.favorite;
+package wooteco.subway.domain.member;
 
 import java.util.Objects;
